// projects/frontend/src/lib/api/models/RegisterDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegisterDto = {
    email: string;
    password: string;
    firstName: string;
    lastName: string;
    telephoneNumber?: string;
    bio?: string;
    birthdate: string;
    sex?: string;
};